refactor sales module
