+++ apps/carbon/app/modules/sales/interfaces/index.ts
@@ -1,3 +1,5 @@
 import useSalesSidebar from "./useSalesSidebar";
 
 export { useSalesSidebar };
+export * from "./CustomerTypes";
+export * from "./Customers";
