+++ apps/carbon/app/routes/api+/sales.customers.tsx
@@ -1,7 +1,7 @@
 import type { LoaderArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
+import { getCustomersList } from "~/modules/sales";
 import { requirePermissions } from "~/services/auth";
-import { getCustomersList } from "~/services/sales";
 
 export async function loader({ request }: LoaderArgs) {
   const authorized = await requirePermissions(request, {
