+++ apps/carbon/app/routes/x+/sales+/customer-types.delete.$customerTypeId.tsx
@@ -2,8 +2,8 @@ import type { ActionArgs, LoaderArgs } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import { useLoaderData, useNavigate, useParams } from "@remix-run/react";
 import { ConfirmDelete } from "~/components/Modals";
+import { deleteCustomerType, getCustomerType } from "~/modules/sales";
 import { requirePermissions } from "~/services/auth";
-import { deleteCustomerType, getCustomerType } from "~/services/sales";
 import { flash } from "~/services/session";
 import { error, success } from "~/utils/result";
 import { notFound } from "~/utils/http";
