+++ apps/carbon/app/routes/x+/sales+/customers.$customerId.location.$customerLocationId.tsx
@@ -1,11 +1,11 @@
 import type { ActionArgs } from "@remix-run/node";
 import { redirect } from "@remix-run/node";
 import { validationError } from "remix-validated-form";
-import { requirePermissions } from "~/services/auth";
 import {
   updateCustomerLocation,
   customerLocationValidator,
-} from "~/services/sales";
+} from "~/modules/sales";
+import { requirePermissions } from "~/services/auth";
 import { flash } from "~/services/session";
 import { assertIsPost, badRequest, notFound } from "~/utils/http";
 import { error, success } from "~/utils/result";
