+++ apps/carbon/app/routes/x+/sales+/customers.tsx
@@ -1,18 +1,15 @@
 import { VStack } from "@chakra-ui/react";
 import type { LoaderArgs } from "@remix-run/node";
-import { redirect } from "@remix-run/node";
-import { json } from "@remix-run/node";
+import { json, redirect } from "@remix-run/node";
 import { Outlet, useLoaderData } from "@remix-run/react";
 import {
   CustomersTable,
   CustomersTableFilters,
-} from "~/interfaces/Sales/Customers";
-import { requirePermissions } from "~/services/auth";
-import {
   getCustomers,
   getCustomerStatuses,
   getCustomerTypes,
-} from "~/services/sales";
+} from "~/modules/sales";
+import { requirePermissions } from "~/services/auth";
 import { flash } from "~/services/session";
 import { getGenericQueryFilters } from "~/utils/query";
 import { error } from "~/utils/result";
